Snoopy: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39260 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/utilities/Snoopy/patches.c
@@ -649,7 +649,7 @@ AROS_LH2(LONG, New_SystemTagList,
     if (patches[PATCH_SystemTagList].enabled)
     {
 	char optstr[20];
-	sprintf(optstr, "%d", result);
+	sprintf(optstr, "%d", (int)result);
 	main_output("SystemTagList", command, optstr, result != -1, TRUE);
     }
